Move table cell properties classes to own namespace.
