+++ src/tablecellproperties/commands/tablecellborderwidthcommand.js
@@ -4,17 +4,17 @@
  */
 
 /**
- * @module table/commands/tablecellborderwidthcommand
+ * @module table/tablecellproperties/commands/tablecellborderwidthcommand
  */
 
 import Command from '@ckeditor/ckeditor5-core/src/command';
 
-import { findAncestor, getSingleValue } from './utils';
+import { findAncestor, getSingleValue } from '../../commands/utils';
 
 /**
  * The table cell border width command.
  *
- * The command is registered by {@link module:table/tablecellpropertiesediting~TableCellPropertiesEditing} as
+ * The command is registered by {@link module:table/tablecellproperties/tablecellpropertiesediting~TableCellPropertiesEditing} as
  * `'tableCellBorderWidth'` editor command.
  *
  * To change cell border width of the selected cell, execute the command:
